Test libGDX PR 7572. WOW!
On the large atlas I tested on, the PR's packer is 10x faster than the old one, give or take 4 milliseconds. 9 seconds with the PR vs. 90 seconds for the old one. The assets I used are in a .7z file because it compresses much more tightly than .zip here. If you extract cg.7z in place, that should get the paths right (it should produce a folder cg/Green/ and a few more, for reference).